requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.12
djangorestframework==3.11.0
idna==2.10
mysqlclient==2.0.1
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
reportlab==3.5.44
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
xlwt==1.3.0
